pyproject.toml
[tool.poetry]
name = "plan-execute-agent-v2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
langchain = "^0.1.17"
langgraph = "^0.0.40"
streamlit = "^1.33.0"
langchain-openai = "^0.1.5"
langsmith = "^0.1.53"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
